Fix sample programs for the mintelf toolchain
This fixes the linker issue:
warning: cannot find entry symbol _start; defaulting to 000000a0

There was an actual problem for gemma.app and test.app.
gcc -O2 puts the main() function into the .text.startup section,
then the linker places it before the custom startup code.
As a result, those programs started directly with main(),
then crashed on exit.

The solution is to explicitly add a global symbol named _start
just before the custom entry point. This way it works in all cases,
even if the entry point isn't at the very start of the program.

Note that the warning was also present when linking kernel32.slb
and gemma32.slb, but there wasn't any actual problem.